Pop completions metatable before registering module functions
luaL_setfuncs and luaL_register don't pop the table they're registering
functions into off the stack, which makes the module functions get
registered into the completions type metatable, and that metatable is
returned from luaopen_linenoise.  This isn't ideal, so we need to pop
the metatable off of the stack after we're done with it.